extract map logic to a component
