+++ modules/map/constants.ts
@@ -1,0 +1,6 @@
+export const MAP_INSETS = {
+  top: 30,
+  right: 0,
+  bottom: 0,
+  left: 10,
+}
